restaurante_api/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
pymongo>=4.6
psycopg2-binary>=2.9
